Move axiomc LSP driver into compiler service boundary
Inline the supported stdio/message loop into axiomc::lsp, remove the separate axiom-lsp workspace crate, and keep readiness regression coverage for both the old Rust-owned driver failure and the migrated driver success.
