--- stage1/crates/axiom-lsp/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "axiom-lsp"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-serde_json.workspace = true
